fix(config): improve configuration merges
The library we were using does not provide a good way to handle
exceptions et al.

Even if it did, we probably need to handle this in a case-by-case basis,
as the correct thing to do for each property is different.

This removes the library, introducing a manual merge process, and starts
adding tests to verify it all.

Ideally, we should also document what happens for each option.

closes #932
closes #241

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>