fix: rethink global config path on windows: use `$HOME/.config`
We made a mistake when it comes to the global config path on Windows.
On macOS and Linux we use two different paths for data and config:
`XDG_DATA_HOME` and `XDG_CONFIG_HOME`. On Windows though, we were using
the same path for both data and config: `LOCALAPPDATA`. This caused
confusion and means the user has no separate dir to store config separate
of data.

In theory we could consider something as `APPDATA`, but honestly, I
propose that we just use `$HOME/.config`. It's easier to find and
consistent with how other tools work. Git and other terminal tools also
use `$HOME/.config`, so users are used to it.

This is not a breaking change because the old path will still load, so
existing setting will load as expected.

Fixes #1347